Code reorg: Move HashedUri into identity_assertion.rs
