+++ src/identity_assertion.rs
@@ -19,7 +19,7 @@ use std::fmt::{Debug, Formatter};
 use serde::{Deserialize, Serialize};
 use serde_bytes::ByteBuf;
 
-use crate::{builder::IdentityAssertionBuilder, c2pa::HashedUri, debug_byte_slice::DebugByteSlice};
+use crate::{builder::IdentityAssertionBuilder, debug_byte_slice::DebugByteSlice};
 
 /// This struct represents the raw content of the identity assertion.
 ///
@@ -166,4 +166,36 @@ impl Debug for IdentityAssertion {
 pub struct SignerPayload {
     /// List of assertions referenced by this credential signature
     pub referenced_assertions: Vec<HashedUri>,
+}
+
+/// A `HashedUri` provides a reference to content available within the same
+/// manifest store.
+///
+/// This is described in §8.3, “[URI References],” of the C2PA Technical
+/// Specification.
+///
+/// [URI References]: https://c2pa.org/specifications/specifications/2.0/specs/C2PA_Specification.html#_uri_references
+#[derive(Clone, Deserialize, Eq, PartialEq, Serialize)]
+pub struct HashedUri {
+    /// JUMBF URI reference
+    pub url: String,
+
+    /// A string identifying the cryptographic hash algorithm used to compute
+    /// the hash
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub alg: Option<String>,
+
+    /// Byte string containing the hash value
+    #[serde(with = "serde_bytes")]
+    pub hash: Vec<u8>,
+}
+
+impl Debug for HashedUri {
+    fn fmt(&self, f: &mut Formatter<'_>) -> Result<(), std::fmt::Error> {
+        f.debug_struct("HashedUri")
+            .field("url", &self.url)
+            .field("alg", &self.alg)
+            .field("hash", &DebugByteSlice(&self.hash))
+            .finish()
+    }
 }
