+++ src/lib.rs
@@ -23,7 +23,7 @@ pub mod c2pa;
 mod debug_byte_slice;
 
 mod identity_assertion;
-pub use identity_assertion::{IdentityAssertion, SignerPayload};
+pub use identity_assertion::{HashedUri, IdentityAssertion, SignerPayload};
 
 #[cfg(test)]
 mod tests;
