+++ src/tests/c2pa/mod.rs
@@ -11,5 +11,4 @@
 // specific language governing permissions and limitations under
 // each license.
 
-mod hashed_uri;
 mod parser;
