+++ src/tests/identity_assertion.rs
@@ -14,7 +14,7 @@
 mod signer_payload {
     use hex_literal::hex;
 
-    use crate::{c2pa::HashedUri, SignerPayload};
+    use crate::{HashedUri, SignerPayload};
 
     #[test]
     fn impl_clone() {
@@ -30,5 +30,33 @@ mod signer_payload {
         };
 
         assert_eq!(signer_payload, signer_payload.clone());
+    }
+}
+
+mod hashed_uri {
+    use hex_literal::hex;
+
+    use crate::HashedUri;
+
+    #[test]
+    fn impl_clone() {
+        let h = HashedUri {
+            url: "self#jumbf=c2pa/urn:uuid:F9168C5E-CEB2-4faa-B6BF-329BF39FA1E4/c2pa.assertions/c2pa.hash.data".to_owned(),
+            alg: Some("sha256".to_owned()),
+            hash: hex!("53d1b2cf4e6d9a97ed9281183fa5d836c32751b9d2fca724b40836befee7d67f").to_vec(),
+        };
+
+        assert_eq!(h.clone(), h);
+    }
+
+    #[test]
+    fn impl_debug() {
+        let h = HashedUri {
+            url: "self#jumbf=c2pa/urn:uuid:F9168C5E-CEB2-4faa-B6BF-329BF39FA1E4/c2pa.assertions/c2pa.hash.data".to_owned(),
+            alg: Some("sha256".to_owned()),
+            hash: hex!("53d1b2cf4e6d9a97ed9281183fa5d836c32751b9d2fca724b40836befee7d67f").to_vec(),
+        };
+
+        assert_eq!(format!("{:#?}", h), "HashedUri {\n    url: \"self#jumbf=c2pa/urn:uuid:F9168C5E-CEB2-4faa-B6BF-329BF39FA1E4/c2pa.assertions/c2pa.hash.data\",\n    alg: Some(\n        \"sha256\",\n    ),\n    hash: 32 bytes starting with [53, d1, b2, cf, 4e, 6d, 9a, 97, ed, 92, 81, 18, 3f, a5, d8, 36, c3, 27, 51, b9],\n}");
     }
 }
